init tests implementation
